src/requirements.txt
flask==2.2.5
mysql-connector-python==8.1.0
prometheus_client==0.15.0